requirements.txt
requests~=2.32.3
pyTelegramBotAPI~=4.23.0
opencv-python~=4.10.0.84
telebot~=0.0.5